Quiet the shell, sort warnings from status, README and issue templates
Warnings and status are captured into shared rings that both front ends show.
The shell no longer prints diagnostics at a player. Link inventories, decode
gaps and PVRTC transcodes are status, not warnings. Version 0.1.0.

CI builds and tests in release, keeps its cache when a job fails, carries a
job timeout, and runs clippy with -D warnings under a documented lint policy.
The toolchain is pinned: with a moving stable, a release adds lints and the
build fails on a machine nobody can reproduce.

CI found what this machine cannot: the cube's per-frame buffers exhausted a
software adapter's mapped-staging arena, a Linux-only test named the NAL
parser by an unresolved path, a VA-API bitfield expectation was two bits off,
and the GPU texture tests ran on a software adapter for over a minute each
instead of skipping. Clippy found a duplicated match arm, a dead match arm, a
getter returning the wrong field's name, an unused signed/unsigned branch in
the ctts parser and a multiply by one.

+++ projects/vitaslop-runtime/src/vita/ctrl.rs
@@ -94,12 +94,12 @@ impl CtrlHistory {
     ///
     /// >>> THE STAMP IS THE VBLANK, NOT THE DISPLAY FRAME, and the pad state is part of the
     /// >>> test. The first version keyed on the display-frame counter, which only the
-    /// preemptive scheduler advances: under the run-to-completion host it never moves, so the
-    /// ring froze on its first (neutral) sample and a title reading more than one buffer never
-    /// saw a button at all. The conformance cube caught it - it pressed START and the guest ran
-    /// on for ever. The vblank counter is a pure function of the clock and advances under every
-    /// host; comparing the pad as well means a change is served the moment the world makes it,
-    /// whatever the clock is doing.
+    /// > > > preemptive scheduler advances: under the run-to-completion host it never moves, so the
+    /// > > > ring froze on its first (neutral) sample and a title reading more than one buffer never
+    /// > > > saw a button at all. The conformance cube caught it - it pressed START and the guest ran
+    /// > > > on for ever. The vblank counter is a pure function of the clock and advances under every
+    /// > > > host; comparing the pad as well means a change is served the moment the world makes it,
+    /// > > > whatever the clock is doing.
     fn push(&mut self, port: u32, vblank: u64, ts: u64, pad: crate::world::CtrlFrame) {
         let ring = match self.ports.iter().position(|(p, _)| *p == port) {
             Some(k) => &mut self.ports[k].1,
@@ -187,7 +187,7 @@ fn fill_ctrl(
                 None => {
                     let base = oldest.map_or((ts, crate::world::CtrlFrame::default()), |s| (s.ts, s.pad));
                     let short = (age + 1 - depth) as u64;
-                    (base.0.saturating_sub(short * u64::from(super::display::VBLANK_US)), base.1)
+                    (base.0.saturating_sub(short * super::display::VBLANK_US), base.1)
                 }
             };
             let bits = if negative { !pad.buttons } else { pad.buttons };
